+++ transport/http/codes_test.go
@@ -49,6 +49,16 @@ func TestUnspecifiedCodes(t *testing.T) {
 		want yarpcerrors.Code
 	}{
 		{
+			name: "code not modified",
+			give: 304,
+			want: yarpcerrors.CodeOK,
+		},
+		{
+			name: "code temporary redirection",
+			give: 307, // test for an x in range: [300, 400)
+			want: yarpcerrors.CodeInvalidArgument,
+		},
+		{
 			name: "code invalid argument",
 			give: 450, // test for an x in range: [400, 500)
 			want: yarpcerrors.CodeInvalidArgument,