+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+- yarpcerrors: classify http 304 as StatusOk and other 3XX statusCode as InvalidArgument.
 
 ## [1.69.1] - 2023-1-24
 ### Changed